# shared/openapi.yaml
openapi: 3.0.3
info:
  title: AgentHub API
  version: 0.1.0
  description: |
    AgentHub 多 Agent 协作平台 API。

    详细规范见 docs/api-spec.md。
  contact:
    name: AgentHub Team

servers:
  - url: http://localhost:8000
    description: Local dev

# ───────────────────────────────────────────────────────
# 安全方案
# ───────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────────────────────
  # 数据模型
  # ─────────────────────────────────────────────────────
  schemas:
    # ─── 通用 ───
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: RESOURCE_NOT_FOUND }
            message: { type: string }
            details: { type: object, additionalProperties: true }

    OffsetPagination:
      type: object
      required: [items, total, page, page_size]
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    CursorPagination:
      type: object
      required: [items, has_more]
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }

    # ─── ContentBlock 联合类型 ───
    TextBlock:
      type: object
      required: [type, text]
      properties:
        type: { type: string, enum: [text] }
        text: { type: string }

    CodeBlock:
      type: object
      required: [type, language, code]
      properties:
        type: { type: string, enum: [code] }
        language: { type: string }
        code: { type: string }

    DiffBlock:
      type: object
      required: [type, filename, before, after]
      properties:
        type: { type: string, enum: [diff] }
        filename: { type: string }
        before: { type: string }
        after: { type: string }

    WebPreviewBlock:
      type: object
      required: [type, url]
      properties:
        type: { type: string, enum: [web_preview] }
        url: { type: string, format: uri }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        thumbnail_url: { type: string, nullable: true }

    FileBlock:
      type: object
      required: [type, filename, url, size, mime_type]
      properties:
        type: { type: string, enum: [file] }
        filename: { type: string }
        url: { type: string, format: uri }
        size: { type: integer }
        mime_type: { type: string }

    ContentBlock:
      oneOf:
        - $ref: '#/components/schemas/TextBlock'
        - $ref: '#/components/schemas/CodeBlock'
        - $ref: '#/components/schemas/DiffBlock'
        - $ref: '#/components/schemas/WebPreviewBlock'
        - $ref: '#/components/schemas/FileBlock'
      discriminator:
        propertyName: type
        mapping:
          text: '#/components/schemas/TextBlock'
          code: '#/components/schemas/CodeBlock'
          diff: '#/components/schemas/DiffBlock'
          web_preview: '#/components/schemas/WebPreviewBlock'
          file: '#/components/schemas/FileBlock'

    # ─── User ───
    User:
      type: object
      required: [id, username, created_at]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        avatar_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_]+$'
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    AuthResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [bearer] }
        expires_in: { type: integer, description: 'seconds' }
        user: { $ref: '#/components/schemas/User' }

    # ─── Conversation ───
    ConversationMode:
      type: string
      enum: [single, group]

    Conversation:
      type: object
      required: [id, title, mode, agent_ids, is_pinned, is_archived, last_message_at, created_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        mode: { $ref: '#/components/schemas/ConversationMode' }
        agent_ids:
          type: array
          items: { type: string }
        is_pinned: { type: boolean }
        is_archived: { type: boolean }
        last_message_at: { type: string, format: date-time }
        last_message_preview: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    CreateConversationRequest:
      type: object
      required: [title, mode, agent_ids]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        mode: { $ref: '#/components/schemas/ConversationMode' }
        agent_ids:
          type: array
          items: { type: string }
          minItems: 1

    UpdateConversationRequest:
      type: object
      properties:
        title: { type: string }
        is_pinned: { type: boolean }
        is_archived: { type: boolean }

    ConversationList:
      allOf:
        - $ref: '#/components/schemas/OffsetPagination'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Conversation' }

    # ─── Message ───
    MessageRole:
      type: string
      enum: [user, agent, system]

    MessageStatus:
      type: string
      enum: [pending, streaming, done, error]

    Message:
      type: object
      required: [id, conversation_id, role, content, status, created_at]
      properties:
        id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/MessageRole' }
        agent_id: { type: string, nullable: true }
        content:
          type: array
          items: { $ref: '#/components/schemas/ContentBlock' }
        reply_to_id: { type: string, format: uuid, nullable: true }
        status: { $ref: '#/components/schemas/MessageStatus' }
        is_pinned: { type: boolean, default: false }
        created_at: { type: string, format: date-time }

    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ContentBlock' }
          minItems: 1
        target_agent_id:
          type: string
          nullable: true
          description: '群聊时指定目标 Agent；单聊可省略'

    SendMessageResponse:
      type: object
      required: [user_message, agent_message]
      properties:
        user_message: { $ref: '#/components/schemas/Message' }
        agent_message: { $ref: '#/components/schemas/Message' }

    UpdateMessageRequest:
      type: object
      properties:
        is_pinned: { type: boolean }

    MessageList:
      allOf:
        - $ref: '#/components/schemas/CursorPagination'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Message' }

    # ─── Agent ───
    AgentProvider:
      type: string
      enum: [claude, openai, custom]

    AgentConfig:
      type: object
      additionalProperties: true
      properties:
        model: { type: string }
        temperature: { type: number, minimum: 0, maximum: 2 }
        max_tokens: { type: integer, minimum: 1, maximum: 16384 }
        top_p: { type: number, minimum: 0, maximum: 1 }

    Agent:
      type: object
      required: [id, name, provider, avatar_url, capabilities, config, is_builtin, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        provider: { $ref: '#/components/schemas/AgentProvider' }
        avatar_url: { type: string }
        capabilities:
          type: array
          items: { type: string }
        system_prompt: { type: string, nullable: true }
        config: { $ref: '#/components/schemas/AgentConfig' }
        is_builtin: { type: boolean }
        created_at: { type: string, format: date-time }

    CreateAgentRequest:
      type: object
      required: [name, provider, config]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        provider: { $ref: '#/components/schemas/AgentProvider' }
        avatar_url: { type: string, default: '' }
        capabilities:
          type: array
          items: { type: string }
          default: []
        system_prompt: { type: string, nullable: true, maxLength: 8192 }
        config: { $ref: '#/components/schemas/AgentConfig' }

    UpdateAgentRequest:
      type: object
      properties:
        name: { type: string }
        avatar_url: { type: string }
        capabilities:
          type: array
          items: { type: string }
        system_prompt: { type: string, nullable: true }
        config: { $ref: '#/components/schemas/AgentConfig' }

    AgentList:
      allOf:
        - $ref: '#/components/schemas/OffsetPagination'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Agent' }

  # ─── 通用响应 ───
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

security:
  - BearerAuth: []

# ───────────────────────────────────────────────────────
# 路径
# ───────────────────────────────────────────────────────
paths:
  # ─── Health ───
  /health:
    get:
      summary: Health check
      security: []
      tags: [Misc]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  # ─── Auth ───
  /api/v1/auth/register:
    post:
      summary: Register a new user
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/auth/login:
    post:
      summary: Login
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/v1/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─── Conversations ───
  /api/v1/conversations:
    get:
      summary: List conversations
      tags: [Conversations]
      parameters:
        - name: archived
          in: query
          schema: { type: boolean, default: false }
        - name: pinned_only
          in: query
          schema: { type: boolean, default: false }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationList' }
    post:
      summary: Create conversation
      tags: [Conversations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateConversationRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/conversations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get conversation
      tags: [Conversations]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update conversation
      tags: [Conversations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateConversationRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }
    delete:
      summary: Delete conversation
      tags: [Conversations]
      responses:
        '204': { description: No Content }

  # ─── Messages ───
  /api/v1/conversations/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: List messages in conversation
      tags: [Messages]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
        - name: direction
          in: query
          schema: { type: string, enum: [before, after], default: before }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageList' }
    post:
      summary: Send message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendMessageRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendMessageResponse' }

  /api/v1/messages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Update message (Pin / Unpin)
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMessageRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
    delete:
      summary: Delete message
      tags: [Messages]
      responses:
        '204': { description: No Content }

  /api/v1/messages/{id}/regenerate:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Regenerate agent message
      tags: [Messages]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }

  /api/v1/messages/{id}/stream:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: SSE stream for agent response
      tags: [Messages]
      description: |
        服务端推送事件流。事件类型见 docs/api-spec.md § 5。
        客户端应使用 fetch-event-source 库（支持自定义 Header）。
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: start
                  data: {"message_id":"...","agent_id":"..."}

                  event: delta
                  data: {"block_index":0,"text_delta":"Hello"}

                  event: done
                  data: {"message_id":"..."}

  # ─── Agents ───
  /api/v1/agents:
    get:
      summary: List agents
      tags: [Agents]
      parameters:
        - name: builtin
          in: query
          schema: { type: boolean }
        - name: provider
          in: query
          schema: { $ref: '#/components/schemas/AgentProvider' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentList' }
    post:
      summary: Create custom agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }

  /api/v1/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get agent
      tags: [Agents]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
    patch:
      summary: Update agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAgentRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
    delete:
      summary: Delete agent
      tags: [Agents]
      responses:
        '204': { description: No Content }